refactor: :arrows_counterclockwise: Update files location for types and styles
