+++ src/screens/Details.tsx
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import { Text, View } from 'react-native';
 import ThemeContext from '../theme/ThemeContext';
-import DetailsStyles from '../styles/Details';
-import { IDetailsProps } from '../types/Details';
-import { IThemeContext } from '../types/ThemeContext';
+import DetailsStyles from '../styles/screens/Details';
+import { IDetailsProps } from '../types/screens/Details';
+import { IThemeContext } from '../types/theme/ThemeContext';
 
 export default class LogIn extends Component<IDetailsProps> {
   static contextType = ThemeContext;
