+++ src/screens/Gps.tsx
@@ -4,9 +4,9 @@ import * as Location from 'expo-location';
 import * as IntentLauncher from 'expo-intent-launcher';
 import * as TaskManager from 'expo-task-manager';
 import ThemeContext from '../theme/ThemeContext';
-import GpsStyles from '../styles/Gps';
-import { IGpsProps, IGpsState } from '../types/Gps';
-import { IThemeContext } from '../types/ThemeContext';
+import GpsStyles from '../styles/screens/Gps';
+import { IGpsProps, IGpsState } from '../types/screens/Gps';
+import { IThemeContext } from '../types/theme/ThemeContext';
 
 export default class Gps extends Component<IGpsProps, IGpsState> {
   static contextType = ThemeContext;
