+++ src/styles/components/CTextInput.ts
@@ -1,6 +1,5 @@
-import { StyleSheet } from 'react-native';
 import CTextInput from '../../components/CTextInput';
-import { IThemeContext } from '../../types/ThemeContext';
+import { IThemeContext } from '../../types/theme/ThemeContext';
 
 export default (context: IThemeContext) => {
   return {
