+++ src/styles/screens/SignUp.ts
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import { IThemeContext } from '../types/ThemeContext';
+import { IThemeContext } from '../../types/theme/ThemeContext';
 
 export default (context: IThemeContext) => {
   return {
