+++ src/types/screens/LogIn.ts
@@ -1,7 +1,5 @@
-import { ContextType } from 'react';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import ThemeContext from '../theme/ThemeContext';
 
-import { RootStackParamList } from './RootStackParamList';
+import { RootStackParamList } from '../navigation/RootStackParamList';
 
 export type ILogInProps = NativeStackScreenProps<RootStackParamList, 'LogIn'>;
