+++ src/types/theme/ThemeContext.ts
@@ -1,4 +1,4 @@
 import { ContextType } from 'react';
-import ThemeContext from '../theme/ThemeContext';
+import ThemeContext from '../../theme/ThemeContext';
 
 export type IThemeContext = ContextType<typeof ThemeContext>;
